OC-310: moved SourceState to top level
